+++ convertkit/includes/blocks/class-convertkit-block-broadcasts.php
@@ -27,9 +27,6 @@ class ConvertKit_Block_Broadcasts extends ConvertKit_Block {
 		// Register this as a Gutenberg block in the ConvertKit Plugin.
 		add_filter( 'convertkit_blocks', array( $this, 'register' ) );
 
-		// Enqueue scripts and styles for this Gutenberg Block in the editor view.
-		add_action( 'convertkit_gutenberg_enqueue_scripts', array( $this, 'enqueue_scripts_editor' ) );
-
 		// Enqueue scripts and styles for this Gutenberg Block in the editor and frontend views.
 		add_action( 'convertkit_gutenberg_enqueue_scripts_editor_and_frontend', array( $this, 'enqueue_scripts' ) );
 		add_action( 'convertkit_gutenberg_enqueue_styles_editor_and_frontend', array( $this, 'enqueue_styles' ) );
@@ -37,17 +34,6 @@ class ConvertKit_Block_Broadcasts extends ConvertKit_Block {
 		// Render Broadcasts block via AJAX.
 		add_action( 'wp_ajax_nopriv_convertkit_broadcasts_render', array( $this, 'render_ajax' ) );
 		add_action( 'wp_ajax_convertkit_broadcasts_render', array( $this, 'render_ajax' ) );
-
-	}
-
-	/**
-	 * Enqueues scripts for this Gutenberg Block in the editor view.
-	 *
-	 * @since   2.0.1
-	 */
-	public function enqueue_scripts_editor() {
-
-		wp_enqueue_script( 'convertkit-gutenberg-block-broadcasts', CONVERTKIT_PLUGIN_URL . 'resources/backend/js/gutenberg-block-broadcasts.js', array( 'convertkit-gutenberg' ), CONVERTKIT_PLUGIN_VERSION, true );
 
 	}
 
@@ -117,11 +103,11 @@ class ConvertKit_Block_Broadcasts extends ConvertKit_Block {
 		$settings = new ConvertKit_Settings();
 
 		return array(
-			'title'                             => __( 'Kit Broadcasts', 'convertkit' ),
-			'description'                       => __( 'Displays a list of your Kit broadcasts.', 'convertkit' ),
-			'icon'                              => 'resources/backend/images/block-icon-broadcasts.svg',
-			'category'                          => 'convertkit',
-			'keywords'                          => array(
+			'title'                         => __( 'Kit Broadcasts', 'convertkit' ),
+			'description'                   => __( 'Displays a list of your Kit broadcasts.', 'convertkit' ),
+			'icon'                          => 'resources/backend/images/block-icon-broadcasts.svg',
+			'category'                      => 'convertkit',
+			'keywords'                      => array(
 				__( 'ConvertKit', 'convertkit' ),
 				__( 'Kit', 'convertkit' ),
 				__( 'Broadcasts', 'convertkit' ),
@@ -129,45 +115,41 @@ class ConvertKit_Block_Broadcasts extends ConvertKit_Block {
 			),
 
 			// Function to call when rendering as a block or a shortcode on the frontend web site.
-			'render_callback'                   => array( $this, 'render' ),
+			'render_callback'               => array( $this, 'render' ),
 
 			// Shortcode: TinyMCE / QuickTags Modal Width and Height.
-			'modal'                             => array(
+			'modal'                         => array(
 				'width'  => 650,
 				'height' => 455,
 			),
 
 			// Shortcode: Include a closing [/shortcode] tag when using TinyMCE or QuickTag Modals.
-			'shortcode_include_closing_tag'     => false,
+			'shortcode_include_closing_tag' => false,
 
 			// Gutenberg: Block Icon in Editor.
-			'gutenberg_icon'                    => convertkit_get_file_contents( CONVERTKIT_PLUGIN_PATH . '/resources/backend/images/block-icon-broadcasts.svg' ),
+			'gutenberg_icon'                => convertkit_get_file_contents( CONVERTKIT_PLUGIN_PATH . '/resources/backend/images/block-icon-broadcasts.svg' ),
 
 			// Gutenberg: Example image showing how this block looks when choosing it in Gutenberg.
-			'gutenberg_example_image'           => CONVERTKIT_PLUGIN_URL . 'resources/backend/images/block-example-broadcasts.png',
+			'gutenberg_example_image'       => CONVERTKIT_PLUGIN_URL . 'resources/backend/images/block-example-broadcasts.png',
 
 			// Help descriptions, displayed when no Access Token / resources exist and this block/shortcode is added.
-			'no_access_token'                   => array(
+			'no_access_token'               => array(
 				'notice'           => __( 'Not connected to Kit.', 'convertkit' ),
 				'link'             => convertkit_get_setup_wizard_plugin_link(),
 				'link_text'        => __( 'Click here to connect your Kit account.', 'convertkit' ),
 				'instruction_text' => __( 'Connect your Kit account at Settings > Kit, and then refresh this page to configure broadcasts to display.', 'convertkit' ),
 			),
-			'no_resources'                      => array(
+			'no_resources'                  => array(
 				'notice'           => __( 'No broadcasts exist in Kit.', 'convertkit' ),
 				'link'             => convertkit_get_new_broadcast_url(),
 				'link_text'        => __( 'Click here to send your first broadcast.', 'convertkit' ),
 				'instruction_text' => __( 'Add a broadcast to your Kit account, and then refresh this page to configure broadcasts to display.', 'convertkit' ),
 			),
 
-			// Gutenberg: JS function to call when rendering the block preview in the Gutenberg editor.
-			// If not defined, render_callback above will be used.
-			'gutenberg_preview_render_callback' => 'convertKitGutenbergBroadcastsBlockRenderPreview',
-
 			// Whether an API Key exists in the Plugin, and are the required resources (broadcasts) available.
 			// If no API Key is specified in the Plugin's settings, render the "No API Key" output.
-			'has_access_token'                  => $settings->has_access_and_refresh_token(),
-			'has_resources'                     => $posts->exist(),
+			'has_access_token'              => $settings->has_access_and_refresh_token(),
+			'has_resources'                 => $posts->exist(),
 		);
 
 	}