+++ convertkit/includes/blocks/class-convertkit-block-product.php
@@ -27,23 +27,9 @@ class ConvertKit_Block_Product extends ConvertKit_Block {
 		// Register this as a Gutenberg block in the ConvertKit Plugin.
 		add_filter( 'convertkit_blocks', array( $this, 'register' ) );
 
-		// Enqueue scripts and styles for this Gutenberg Block in the editor view.
-		add_action( 'convertkit_gutenberg_enqueue_scripts', array( $this, 'enqueue_scripts_editor' ) );
-
 		// Enqueue scripts and styles for this Gutenberg Block in the editor and frontend views.
 		add_action( 'convertkit_gutenberg_enqueue_scripts_editor_and_frontend', array( $this, 'enqueue_scripts' ) );
 		add_action( 'convertkit_gutenberg_enqueue_styles_editor_and_frontend', array( $this, 'enqueue_styles' ) );
-
-	}
-
-	/**
-	 * Enqueues scripts for this Gutenberg Block in the editor view.
-	 *
-	 * @since   1.9.8.5
-	 */
-	public function enqueue_scripts_editor() {
-
-		wp_enqueue_script( 'convertkit-gutenberg-block-product', CONVERTKIT_PLUGIN_URL . 'resources/backend/js/gutenberg-block-product.js', array( 'convertkit-gutenberg' ), CONVERTKIT_PLUGIN_VERSION, true );
 
 	}
 
@@ -109,59 +95,56 @@ class ConvertKit_Block_Product extends ConvertKit_Block {
 		$settings            = new ConvertKit_Settings();
 
 		return array(
-			'title'                             => __( 'Kit Product', 'convertkit' ),
-			'description'                       => __( 'Displays a button to purchase a Kit product.', 'convertkit' ),
-			'icon'                              => 'resources/backend/images/block-icon-product.svg',
-			'category'                          => 'convertkit',
-			'keywords'                          => array(
+			'title'                                => __( 'Kit Product', 'convertkit' ),
+			'description'                          => __( 'Displays a button to purchase a Kit product.', 'convertkit' ),
+			'icon'                                 => 'resources/backend/images/block-icon-product.svg',
+			'category'                             => 'convertkit',
+			'keywords'                             => array(
 				__( 'ConvertKit', 'convertkit' ),
 				__( 'Kit', 'convertkit' ),
 				__( 'Product', 'convertkit' ),
 			),
 
 			// Function to call when rendering as a block or a shortcode on the frontend web site.
-			'render_callback'                   => array( $this, 'render' ),
+			'render_callback'                      => array( $this, 'render' ),
 
 			// Shortcode: TinyMCE / QuickTags Modal Width and Height.
-			'modal'                             => array(
+			'modal'                                => array(
 				'width'  => 600,
 				'height' => 518,
 			),
 
 			// Shortcode: Include a closing [/shortcode] tag when using TinyMCE or QuickTag Modals.
-			'shortcode_include_closing_tag'     => false,
+			'shortcode_include_closing_tag'        => false,
 
 			// Gutenberg: Block Icon in Editor.
-			'gutenberg_icon'                    => convertkit_get_file_contents( CONVERTKIT_PLUGIN_PATH . '/resources/backend/images/block-icon-product.svg' ),
+			'gutenberg_icon'                       => convertkit_get_file_contents( CONVERTKIT_PLUGIN_PATH . '/resources/backend/images/block-icon-product.svg' ),
 
 			// Gutenberg: Example image showing how this block looks when choosing it in Gutenberg.
-			'gutenberg_example_image'           => CONVERTKIT_PLUGIN_URL . 'resources/backend/images/block-example-product.png',
+			'gutenberg_example_image'              => CONVERTKIT_PLUGIN_URL . 'resources/backend/images/block-example-product.png',
 
 			// Help descriptions, displayed when no Access Token / resources exist and this block/shortcode is added.
-			'no_access_token'                   => array(
+			'no_access_token'                      => array(
 				'notice'           => __( 'Not connected to Kit.', 'convertkit' ),
 				'link'             => convertkit_get_setup_wizard_plugin_link(),
 				'link_text'        => __( 'Click here to connect your Kit account.', 'convertkit' ),
 				'instruction_text' => __( 'Connect your Kit account at Settings > Kit, and then refresh this page to select a product.', 'convertkit' ),
 			),
-			'no_resources'                      => array(
+			'no_resources'                         => array(
 				'notice'           => __( 'No products exist in Kit.', 'convertkit' ),
 				'link'             => convertkit_get_new_product_url(),
 				'link_text'        => __( 'Click here to create your first product.', 'convertkit' ),
 				'instruction_text' => __( 'Add a product to your Kit account, and then refresh this page to select a product.', 'convertkit' ),
 			),
 
-			// Gutenberg: Help descriptions, displayed when no settings defined for a newly added Block.
-			'gutenberg_help_description'        => __( 'Select a Product using the Product option in the Gutenberg sidebar.', 'convertkit' ),
-
-			// Gutenberg: JS function to call when rendering the block preview in the Gutenberg editor.
-			// If not defined, render_callback above will be used.
-			'gutenberg_preview_render_callback' => 'convertKitGutenbergProductBlockRenderPreview',
+			// Gutenberg: Help description, displayed when the defined attribute has not been set for the block.
+			'gutenberg_help_description'           => __( 'Select a Product using the Product option in the Gutenberg sidebar.', 'convertkit' ),
+			'gutenberg_help_description_attribute' => 'product',
 
 			// Whether an Access Token exists in the Plugin, and are the required resources (products) available.
 			// If no Access Token is specified in the Plugin's settings, render the "Not Connected" output.
-			'has_access_token'                  => $settings->has_access_and_refresh_token(),
-			'has_resources'                     => $convertkit_products->exist(),
+			'has_access_token'                     => $settings->has_access_and_refresh_token(),
+			'has_resources'                        => $convertkit_products->exist(),
 		);
 
 	}